More cleanup - remove Waveform class and move Queue to Utils
